separate new types into 'namespaces'
