+++ src/Dashboards/index.ts
@@ -1,0 +1,1 @@
+export * as WidgetConfigHelpers from "./WidgetConfigHelpers";
